+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 
 ### Added
 
+### Changed
+
+### Fixed
+
+### Breaking Change
+
+
+## 0.3.0
+
+### Added
+
 - feat(template): Add new filter 'format_column_name' to beautify the column name for tables
 - feat(template.Service): Display service status
 - feat(view.Service): Allow to update service status